Troubleshoot selectedItemReducer bug.

+++ src/components/ItemDetails.js
@@ -1,26 +1,28 @@
 import React from 'react'
 import { connect } from 'react-redux';
-import { editItem } from './../actions';
+import { updateItem, deleteItem } from './../actions';
 
 
 function ItemDetails(props) {
   const { item } = props;
 
-  return (
-    <div className="ItemDetails" key={item.id}>
-      {item.itemName}:  ${item.price}
-      {/* <button onClick={() => props.deleteItem(item)}>Purchase</button> */}  
-      <button onClick={() => props.editItem(item)}>Edit</button>
-      <button onClick={() => props.deleteItem(item)}>Delete</button>
-    </div>
-  )
+  if (props.selectedItem) {
+    return (
+      <div className="ItemDetails" key={item.id}>
+        {item.itemName}:  ${item.price}
+        {/* <button onClick={() => props.deleteItem(item)}>Purchase</button> */}  
+        <button onClick={() => props.updateItem(item)}>Edit</button>
+        <button onClick={() => props.deleteItem(item)}>Delete</button>
+      </div>
+    )
+  }
 }
 
 const mapStateToProps = state => {
   return {
     selectedItem: state.selectedItemReducer.selectedItem,
-    itemToEdit: state.selectedItemReducer.itemToEdit
+    editing: state.selectedItemReducer.editing
   }
 }
 
-export default connect(mapStateToProps, { editItem })(ItemDetails);
+export default connect(mapStateToProps, { updateItem, deleteItem })(ItemDetails);
